requirements.txt
pandas
requests
matplotlib
prophet
numpy
scikit-learn
tensorflow-cpu
